double handling error, remove array testing

+++ src/java.base/share/classes/java/lang/runtime/Carriers.java
@@ -348,7 +348,7 @@ public final class Carriers {
              * @param i  array index
              */
             private long getLong(int i) {
-                return UNSAFE.getInt(primitives, offsetToLong(i));
+                return UNSAFE.getLong(primitives, offsetToLong(i));
             }
 
             /**
